@@ -7,7 +7,7 @@
   "homepage": "https://www.makingmorehealth.org/",
   "license": "GPL-3.0-or-later",
   "scripts": {
-    "build": "react-scripts build",
+    "build": "vite build",
     "build:ionic:browser": "ionic build --prod --engine=browser",
     "build:ionic:ios": "ionic capacitor sync ios --prod",
     "eject": "react-scripts eject",
@@ -21,26 +21,25 @@
     "release:github": "semantic-release --extends './scripts/release/releaserc.github.js'",
     "release:information": "semantic-release --dry-run --extends './scripts/release/releaserc.information.js' --no-ci",
     "repair": "npx rimraf build node_modules .git/hooks package-lock.json && npm install && npx husky install",
-    "start": "react-scripts start",
+    "start": "vite",
     "start:prod": "npx http-server ./build --proxy http://localhost:8080? --silent",
     "stylelint-check": "stylelint-config-prettier-check",
-    "test": "react-scripts test"
+    "test": "react-scripts test --transformIgnorePatterns 'node_modules/(?!(@ionic/react|@ionic/react-router|@ionic/core|@stencil/core|ionicons)/)'",
+    "serve": "vite preview"
   },
   "dependencies": {
-    "@capacitor/android": "^3.0.0-beta.1",
-    "@capacitor/core": "^3.0.0-beta.1",
+    "@capacitor/android": "3.0.0-beta.1",
+    "@capacitor/core": "3.0.0-beta.1",
     "@capacitor/filesystem": "^0.5.6",
-    "@capacitor/ios": "^3.0.0-beta.1",
-    "@capacitor/splash-screen": "^0.3.10",
-    "@capacitor/status-bar": "^0.4.10",
-    "@capacitor/storage": "^0.3.10",
-    "@ionic/react": "^5.6.6",
-    "@ionic/react-router": "^5.6.6",
+    "@capacitor/ios": "3.0.0-beta.1",
+    "@capacitor/splash-screen": "0.3.10",
+    "@capacitor/status-bar": "0.4.10",
+    "@capacitor/storage": "0.3.10",
+    "@ionic/react": "^6.7.5",
+    "@ionic/react-router": "^6.7.5",
     "@ionic/storage": "^3.0.4",
     "@reduxjs/toolkit": "^1.5.1",
-    "@testing-library/jest-dom": "5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^13.1.9",
+    "app-provider": "^1.1.2",
     "fuse.js": "^6.4.6",
     "i18next": "^20.2.2",
     "i18next-browser-languagedetector": "6.1.1",
@@ -48,15 +47,12 @@
     "md5": "^2.3.0",
     "microsoft-cognitiveservices-speech-sdk": "^1.16.0",
     "query-string": "^7.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-i18next": "^11.8.15",
-    "react-redux": "7.2.4",
-    "react-router": "5.2.0",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.0",
-    "redux-persist": "^6.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^13.0.3",
+    "react-redux": "8.0.5",
+    "react-router-dom": "5.3.4",
+    "redux-persist": "6.0.0",
     "typescript": "^4.2.4",
     "use-query-params": "^1.2.2"
   },
@@ -77,6 +73,7 @@
     "@types/react-redux": "^7.1.16",
     "@types/react-router": "^5.1.14",
     "@types/react-router-dom": "5.1.7",
+    "@vitejs/plugin-react": "^4.0.3",
     "chalk": "^4.1.1",
     "device-bars": "^1.1.2",
     "eslint-config-prettier": "^8.3.0",
@@ -96,7 +93,8 @@
     "stylelint-config-standard": "^22.0.0",
     "stylelint-order": "^4.1.0",
     "stylelint-prettier": "^1.2.0",
-    "ts-jest": "^26.5.6"
+    "ts-jest": "^26.5.6",
+    "vite": "^4.4.7"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
